Fix fast build (#60)
* Revert "config to switch linker to mold for linux target (#51)"

This reverts commit 860547f453955857a4ff0160008df446c2cf579c.

PROBLEM: messes up docker push
SOLUTION: remove linker switch

NOTE:
issues mentioned in 86054 still apply, and will be addressed
in this PR.

* add fast_build script that runs cargo build with the mold linker

PROBLEM:
as mentioned in 86054, hot-reloads take around 10s, mostly due to
linker.

SOLUTION:
use mold, this time via a fast_build script that sets the necessary
cargo env vars, ensures that mold and clang are installed, and runs
cargo build with the correct path for the mold linker.

POSSIBLE IMPROVEMENTS:
may be a good idea to switch the shell script for a zx script. Also
would be nice if `cargo build` automatically ran fast_build.sh as a
wrapper if the host and profile are correct (ie. linux host, with
dev profile)

* CLEANUP: remove .cargo from dockerignore

Co-authored-by: Abhishek Cherath <abhi@driftingin.space>