remove recv.close()
this was only necessary when the if condition only checked for Swept,
since in that case the self.step(req, BackendState::Terminated) would
cause a Signal::Interrupt before returning but after killing the backend,
which got it back to the select! and ran the termination twice.

Since we ignore the signal handler in a terminal state now, there is no
chance of that happening.