Abhi/fix no sweeps (#390)
* fixes issue where backend destructor doesn't run

I misunderstood how the destructor code gets called. It gets called in the transition for
a terminal state, not TO a terminal state.

* use old logic

* remove recv.close()

this was only necessary when the if condition only checked for Swept,
since in that case the self.step(req, BackendState::Terminated) would
cause a Signal::Interrupt before returning but after killing the backend,
which got it back to the select! and ran the termination twice.

Since we ignore the signal handler in a terminal state now, there is no
chance of that happening.