apparently awaiting while holding a lock is highly verboten
https://stackoverflow.com/questions/7612602/why-cant-i-use-the-await-operator-within-the-body-of-a-lock-statement

"Correct, you have discovered why we made it illegal.
Awaiting inside a lock is a recipe for producing deadlocks.

I'm sure you can see why: arbitrary code runs between the time the
await returns control to the caller and the method resumes.
That arbitrary code could be taking out locks that produce
lock ordering inversions, and therefore deadlocks."

further, *it is impossible to lock a mutex in one thread and unlock it
in another*, this means that it's not Send, which means the wait_for_seq
function cannot be called in tokio::Spawn.

fortunately if we stick the listener in a dashmap we can mess with it
from a RwLockReadGuard, so it's nbd.