emit lock removal messages
NOTE: not sure about this, is it possible for the state.state() read
lock to give a state a few time units before these messages?

in that case, it may be possible that some locks do not get their
remove messages sent.  (TOCTOU)

Ideally the lock removal messages would be emitted by the state
machine itself.

Although I guess it may be possible that having a message to remove a
lock is not the right approach. since it leads to logical_times where
there is a lock pointing to some backend, but the backend is terminal.

this could be circumvented by introducing transactions, but in
conclusion; not the right approach for the time being.