fix lock acquisition race condition (#427)
Adds lock_assignment field to BackendAssignment message, this is of forme BackendLockAssignment, which has two fields, a lock string, and a uid that serves as a fencing token. 

The uid ensures that a lock is only ever acquired by one backend. It is obtained in the  ClusterStateMessage::LockMessage variant Announce, which is sent by the controller prior to submitting a spawn request and backend assignment.

This PR also adds a ClusterStateMessage::LockMessage variant Revoke in order to manually remove a lock if needed.

This fixes the race condition bug where multiple backends could acquire the same lock if spawn requests were simultaneous or right after one another.

Co-authored-by: Paul Butler <paul@driftingin.space>