Abhi/dylib speedup tests (#519)
This creates a sub crate for tests, and has that crate use a plane 2 dylib. The only thing that really changes for users is now you've to run cargo test --all to run the integration tests.

recompiling tests after adding a println or something now takes 10 seconds and ~200mb! which is still a bit long, but it's a lot better than 3 minutes and ~8gb earlier.