Mark backend as terminated when lost (#746)
In the weird case where the Plane controller asks a drone to terminate a
backend that has already been removed from Docker, we currently do
nothing but don't do anything to tell the controller to mark the backend
as lost. This changes the behavior to do that, via the state store.