Add last_status_number to backend table (#756)
This adds a (nullable) `last_status_number` column to backend. A
follow-up PR will populate this column.

Per [postgres
docs](https://www.postgresql.org/docs/current/sql-altertable.html#SQL-ALTERTABLE-NOTES):

> When a column is added with ADD COLUMN and a non-volatile DEFAULT is
specified, the default is evaluated at the time of the statement and the
result stored in the table's metadata. That value will be used for the
column for all existing rows. If no DEFAULT is specified, NULL is used.
In neither case is a rewrite of the table required.

So this is a cheap migration.