Emit scheduled event in connect transaction (#761)
This refactors the code that:
- updates the `backend_state` table
- emits the `BackendState` event

Into one helper function that takes a transaction to execute in. This
code path is then used both when updating state on the drone, and when
creating the initial `Scheduled` state on the server.

The `last_status_time`, `backend_state.created_at`, and the event's
timestamp will all be the time of the transaction start.