Make ready address required (#776)
It doesn't make sense to have a ready backend without an address. This
avoids that situation by passing the address through `to_ready`.

`BackendState` is sent across the network, so changes to it need to be
version-safe. This change is version-safe:

- If a **writer** is on this version and a **reader** is on a prior
version, nothing bad can happen. The reader continues to accept `null` a
the value, but the writer will never produce that value.
- If a **reader** is on this version and a **writer** is on a prior
version, there is technically a code path that produces a `null` here,
but:
  - It's an error code path, and it logs a warning message
  - I could not find any instances of this in our logs