Introduce SocketPath type with copy-safe drop (#779)
We want `server` to be copy safe: we do this successfully by wrapping
the `_loop_task` in an `Arc` `GuardHandle`. However, `server`'s `Drop`
will attempt to delete the `socket_path` file on every dropped copy --
which isn't the behavior we want. We only want that file deleted when
the last copy is dropped. We achieve this similar to `_loop_task`: by
creating a new `SocketPath` type (with its own `Drop` impl) and wrapping
this type in an `Arc`.