Expose docker utils for re-use (#788)
Two things:
1. Allow me to access `docker` under the hood of the `DockerRuntime` to
do custom stuff in a separate runtime. I still want to keep
`DockerRuntime` around because some functionality is identical and it
allows me to lean on the existing runtime's implementation of stuff.
2. Allow me to re-use the container config, which is by and large
identical in a separate runtime -- except for the `backend_name`. Turns
out it's not that hard to make that optional!