Split `Terminating` into (soft) `Terminating` and `HardTerminating` (#791)
Currently, we use `Terminating` to represent both soft and hard
terminating. If a backend is hard-terminated while it is being
soft-terminated (which is allowed), a backend goes from a `Terminating`
state to another `Terminating` state. This is the only allowed state
transition that does not move forward in the state sequence.

The change in this PR is to split soft- and hard- `Terminating` states
into their own separate states. Soft-termination keeps the name
`Terminating`, and hard termination gets `HardTerminating`. The field
`termination` in `Terminating` remains, but is deprecated, and can only
be set to `Soft` in the constructor.