Refactor proxy (#814)
This refactors the reverse proxy embedded in Plane.

The main goal of this is to update hyper support to the `1.x` branch. As
side-goals, it:
- Refactors the proxy substantially to separate proxy logic from Plane
business logic
- Adds a bunch of tests and some testing infrastructure

Rather than upgrading the entire `plane` project to the latest version
of hyper, this PR introduces a `dynamic-proxy` crate which contains the
HTTP-level proxy implementation using the latest stable version of each
of its dependencies. This allows us to decouple the dependency versions
of the proxy from the rest of Plane.

Once this is merged, the next step will be to update the `axum` and
`reqwests` crates in `plane` itself so that they use the latest `hyper`.
Then, we can either keep `dynamic-proxy` as a separate crate and publish
it (e.g. as `plane-dynamic-proxy`), or merge it back in to the main
`plane` crate.

[Requirements
matrix](https://docs.google.com/document/d/1cxc-8xamW6BzoflsQmKkW8pwKm8USVD7Tq1PZnpoR_4/edit)