Add connection timeout (#837)
We used to have a timeout on the lifespan of the entire request, which
also captured an initial connection timeout, but also failed for
long-lived connections like large requests. When we removed that
timeout, we also removed the ability to time-out a request when the
initial TCP connection hangs.

This adds back a timeout that only applies to the initial connection.