drone failed to mark backend as terminated (#838)
We've observed failure modes where a container exit is not detected. In
those cases, we expect that a termination request will see that the
container is no longer running due to docker returning an error when
calling `stop_container`/`kill_container`. Docker _doesn't_ error when
calling `stop_container()` on an already stopped container, however.
This has caused termination loops that never fully succeed, leaving the
drone thinking the container is still running, and re-renewing its key
over and over. This PR fixes this issue by checking the container's
status before calling `terminate()` on it.

This still doesn't address these two issues, however:
(1) we aren't detecting the container exit in some cases which causes
this failure mode in the first place
(2) we should make `terminating` loops eventually resort to
`hard-terminating`