Refactor parts of Plane into a `plane-common` crate (DIS-2967) (#849)
The main `plane` crate contains a client, but depending on it in another
project requires bringing in _all_ of Plane, including the openssl
dependency. This breaks the client (and other pieces) out into a
`plane-common` crate with the parts needed to "consume" Plane as a
client, and a crate for Plane itself (in particular, without the
database and ACME stuff, which bring in other dependencies that can be
hard to wrangle).

This looks like a huge PR but there is essentially no net new code here;
files are just moved around. A few notes:
- I got rid of `plane-dynamic`; it was a hack to make tests compile
faster but it should be less necessary now that we have broken up Plane
into several crates.
- Roughly, the criteria for "what goes in common" is "the client and
everything it depends on". `plane` depends on `plane-common`, but
`plane-common` does NOT depend on `plane`, so things like
`ExponentialBackoff` (used by the client to manage reconnects) go in
`plane-common`.

Todo:
- [x] get `test_get_metrics` and `backend_lifecycle` tests to pass.
- [x] rename to `plane-common`?
- [x] p2 PR for refactor and ensure p2 integration tests pass