Don't break build when git is missing (#856)
If git is missing, the build script will fail. This changes it so that
it uses the string `unknown` instead of the version.

This version is currently only used to log a warning if there is a
version mismatch between the client and server. We should switch to
using only the version number for this soon anyway now that the protocol
is more stable.