Use `UnsyncBoxBody` instead of `BoxBody` (#860)
We currently use `BoxBody` extensively, which [requires the body to be
`Sync`](https://docs.rs/http-body-util/latest/http_body_util/combinators/struct.BoxBody.html),
even though we don't actually rely on the `Body` being `Sync` in our
code.

[Axum uses
`UnsyncBoxBody`](https://docs.rs/axum-core/0.5.0/src/axum_core/body.rs.html#14)
instead, which is like `BoxBody` but allows the contained body to be
`!Sync`. So when we get a `Body` from Axum, we can't currently pass it
to `dynamic_proxy` because it may not be `Sync`, even though our code
doesn't actually care if it is `Sync`.

This replaces our use of `BoxBody` with `UnsyncBoxBody`.