Remove backends from connection monitor (#868)
The connection monitor works by visiting backends from a queue. Each
queue entry points to a map entry that is currently always expected to
exist as an invariant. If it ever is not present, however, we (already)
skip it and do _not_ insert the value back in the queue.

This PR piggybacks on this existing behavior to remove backends from
being cycled through the queue repeatedly. When we remove a backend from
the route map we also remove it from the monitor map, which will cause
it to be removed from the queue once it reaches the top.